@@ -1,10 +1,13 @@
 import { Router } from 'express';
-import { subscriptionChargeFailedNotification } from '../sendgrid_mailer';
-var moment = require('moment');
-import db from '../../models';
 
+import { subscriptionChargeFailedNotification } from '../sendgrid_mailer';
+import db from '../../models';
 import stripe from '../stripe';
-var async = require('async');
+import invoicePaymentSucceeded from '../../utils/invoice_payment_succeeded_webhook';
+import chargeSucceededWebhook from '../../utils/charge_succeeded_webhook';
+
+const moment = require('moment');
+const async = require('async');
 
 const router = new Router({ mergeParams: true });
 
@@ -26,7 +29,6 @@ function waterfaller(functions) {
 }
 
 function stripe_webhook(request, response) {
-
   var { body } = request;
   console.log(`webhook event: ${body.type}`);
 
@@ -40,121 +42,7 @@ function stripe_webhook(request, response) {
   }
 
   if (body.type === 'charge.succeeded') {
-    function queryPaymentProfile(callback) {
-      let stripeCustomerId = body.data.object.customer;
-      db.PaymentProfile.findOne({
-        where: {
-          stripeCustomerId
-        }
-      }).then(paymentProfile => {
-        if (!paymentProfile) {
-          return callback('No matching records in payment profile');
-        }
-        let paymentProfileId = paymentProfile.id;
-        callback(null, paymentProfileId);
-      });
-    }
-
-    function getClientSubscriptions(paymentProfileId, callback) {
-      db.Subscription.findAll({
-        where: {
-          paymentProfileId
-        }
-      }).then(clientSubscriptions => {
-        callback(null, clientSubscriptions);
-      })
-    }
-
-    function getDentistMembershipPlans(clientSubscriptions, callback) {
-      if (clientSubscriptions.length > 0) {
-        let { dentistId } = clientSubscriptions[0];
-        db.Membership.findAll({
-          where: {
-            userId: dentistId
-          }
-        }).then(dentistMembershipPlans => {
-          callback(null, clientSubscriptions, dentistMembershipPlans);
-        })
-      }
-      else {
-        callback(null, [], []);
-      }
-    }
-
-    function checkAdultUsers(clientSubscriptions = [], dentistMembershipPlans = [], callback) {
-      let thirteen_years_ago = moment().subtract("13", "years").add("1", "month").format("YYYY-MM-DD");
-      let clientIds = clientSubscriptions.map(subscription => subscription.clientId);
-      if (clientIds.length > 0) {
-        db.User.findAll({
-          where: {
-            id: {
-              $in: clientIds
-            },
-            birthDate: {
-              $lte: thirteen_years_ago
-            }
-          }
-        }).then(clients => {
-          let matchingClientIds = clients.map(client => client.id);
-          clientSubscriptions = clientSubscriptions.filter(sub => matchingClientIds.includes(sub.clientId));
-          callback(null, clientSubscriptions, dentistMembershipPlans);
-        });
-      }
-      else {
-        callback(null, clientSubscriptions, dentistMembershipPlans)
-      }
-    }
-
-    function checkAndUpdateChildMembershipPlans(clientSubscriptions = [], dentistMembershipPlans = [], callback) {
-
-      if (clientSubscriptions.length > 0 && dentistMembershipPlans.length > 0) {
-        async.each(clientSubscriptions, (subs, eachCallback) => {
-          let clientPlan = dentistMembershipPlans.find(plan => plan.id == subs.membershipId);
-          let dentistChildMemberShip = dentistMembershipPlans.find(plan => plan.name == 'default child membership' && plan.type == "month");
-          let dentistAdultMemberShip = dentistMembershipPlans.find(plan => plan.name == 'default membership' && plan.type == "month");
-          if (clientPlan.name !== 'default membership' && clientPlan.type == "month" && (subs.status == "active" || subs.status == "past_due")) {
-            stripe.updateSubscription(subs.stripeSubscriptionId, dentistAdultMemberShip.stripePlanId, true)
-              .then(data => {
-                subs.membershipId = dentistAdultMemberShip.id;
-                if (subs.status == "past_due") {
-                  subs.status = "active";
-                }
-                subs.save();
-                eachCallback();
-              }, (err) => {
-                eachCallback(err);
-              });
-          }
-          else {
-            eachCallback();
-          }
-        }, (err) => {
-          if (err) {
-            callback(err);
-          }
-          else {
-            callback(null, "charge_succeeded hook executed Successfully.");
-          }
-        });
-      }
-      else {
-        callback(null, "charge_succeeded hook executed Successfully with NO changes.");
-      }
-    }
-
-    async.waterfall([
-      queryPaymentProfile,
-      getClientSubscriptions,
-      getDentistMembershipPlans,
-      checkAdultUsers,
-      checkAndUpdateChildMembershipPlans
-    ], (err, result) => {
-      if (err) {
-        console.log(err);
-      } else {
-        console.log(result);
-      }
-    });
+    chargeSucceededWebhook(body);
   }
 
   else if (body.type == "charge.failed") {
@@ -338,6 +226,11 @@ function stripe_webhook(request, response) {
       ]
     ).then(data => console.log("invoice.created hook executed successfully"), err => console.log(err));
   }
+
+  else if (body.type == "invoice.payment_succeeded") {
+    invoicePaymentSucceeded(body);
+  }
+  
   response.status(200).send({});
 }
 
