@@ -101,7 +101,7 @@ export default {
   dentistReviewNotification(res, user, patient, review) {
     return mail(res.mailer, 'dentists/new_review', {
       to: user.email,
-      subs: EMAIL_SUBJECTS.dentist.new_review,
+      subject: EMAIL_SUBJECTS.dentist.new_review,
       site,
       patient,
       review
@@ -114,10 +114,10 @@ export default {
 function constructDentistInfo(dentist) {
   let { dentistInfo } = dentist;
   return `
-    Dentist Name: ${dentist.firstName + dentist.lastName}\n
-    Office Name: ${dentistInfo.officeName}\n
-    Dentist Email: ${dentist.email}\n
-    Dentist Office Email: ${dentistInfo.email}\n
+    Dentist Name: ${dentist.firstName + dentist.lastName},
+    Office Name: ${dentistInfo.officeName},
+    Dentist Email: ${dentist.email},
+    Dentist Office Email: ${dentistInfo.email},
     Office Address: ${dentistInfo.address || ''}, ${dentistInfo.city || ''}, ${dentistInfo.state || ''}, ${dentistInfo.zipcode || ''}\n
   `.replace(/  /g,'');
 }
@@ -133,7 +133,7 @@ function constructPaymentDetails(subs, plans) {
   },0);
 
   return `
-    Total Family Members Subscribed: ${subs.length}\n
+    Total Family Members Subscribed: ${subs.length},
     Subtotal: ${String(total)}.00 $
   `.replace(/  /g,'');
 }
