@@ -17,6 +17,8 @@ import {
   BadRequestError,
 } from '../errors';
 
+import { processDiff } from '../../utils/compareUtils';
+import { notifyPlanUpdate } from '../../helpers/membership';
 // ────────────────────────────────────────────────────────────────────────────────
 // ROUTER
 
@@ -65,17 +67,42 @@ function getDentistInfo(req, res, next) {
     dentistInfo.activeMemberCount = activeMemberCount;
 
     if (req.user.get('type') === 'dentist') {
-      res.json({
+      req.dentistInfoResult = {
         data: dentistInfo,
         stripe_public_key: process.env.STRIPE_PUBLIC_KEY
-      });
+      };
+      next();
     } else {
       const data = req.user.toJSON();
       data.dentistInfo = dentistInfo;
-      res.json({ data, stripe_public_key: process.env.STRIPE_PUBLIC_KEY });
+      req.dentistInfoResult = {
+        data,
+        stripe_public_key: process.env.STRIPE_PUBLIC_KEY
+      };
+      next();
     }
   })
   .catch(err => next(new BadRequestError(err)));
+}
+
+async function getCustomMembership (req, res) {
+  const dentistInfo = req.dentistInfoResult.data;
+  if (req.query.custom_plans) {
+    const custom_memberships = await db.Membership.findAll({
+      where: {
+        dentistInfoId: dentistInfo.id,
+        active: true,
+        type: 'custom'
+      },
+      include: [{
+        model: db.CustomMembershipItem,
+        as: 'custom_items'
+      }]
+    }).map(m => m.toJSON());
+    req.dentistInfoResult.data.custom_memberships = custom_memberships;
+  }
+  
+  return res.status(200).send({ ...req.dentistInfoResult });
 }
 
 /**
@@ -86,225 +113,232 @@ function getDentistInfo(req, res, next) {
  * @param {Function} next - the next middleware function
  */
 async function updateDentistInfo(req, res, next) {
-  const userId = req.params.dentistInfoId;
+  const userId = req.params.dentistId;
 
-  /*
-   * If user is not admin and try to requests paths not related
-   * to that user will return forbidden.
-   */
-  const canEdit =
-    userId === 'me' || req.user.get('id') === parseInt(userId, 10) ||
-    (req.user.get('type') === 'admin' && userId !== 'me');
+  const { dentistInfo } = req.locals;
+  const queries = [];
+  const user = req.body.user;
+  const officeInfo = req.body.officeInfo;
+  const officeImages = req.body.officeImages;
+  const pricing = req.body.pricing;
+  const membership = req.body.officeInfo.memberships;
+  const childMembership = req.body.childMembership;
+  const workingHours = req.body.workingHours;
+  const services = req.body.services;
 
-  if (!canEdit) {
-    return next(new ForbiddenError());
+  let shouldRefresh = false;
+
+  if (user) {
+    queries.push(
+      req.user.update({
+        firstName: user.firstName,
+        lastName: user.lastName,
+        phone: user.phone,
+        avatar: user.avatar,
+        zipCode: user.zipCode,
+        dentistSpecialtyId: user.dentistSpecialtyId
+      })
+    );
   }
 
-  const query = {
-    where: {},
-    include: [{
-      model: db.DentistInfoPhotos,
-      as: 'officeImages'
-    },
-    {
-      model: db.DentistInfoService,
-      as: 'services'
-    }]
-  };
-
-  if (req.params.dentistInfoId) {
-    query.where.userId = req.params.dentistInfoId;
+  // update info itself.
+  if (officeInfo) {
+    queries.push(
+      dentistInfo.update({
+        officeName: officeInfo.officeName,
+        url: officeInfo.url,
+        phone: officeInfo.phone,
+        message: officeInfo.message,
+        address: officeInfo.address,
+        city: officeInfo.city,
+        state: officeInfo.state,
+        zipCode: officeInfo.zipCode,
+        logo: officeInfo.logo,
+        acceptsChildren: officeInfo.acceptsChildren,
+        childStartingAge: officeInfo.childStartingAge,
+        marketplaceOptIn: officeInfo.marketplaceOptIn
+      })
+    );
   }
 
-  // if not admin limit query to related data userId
-  if (req.user.get('type') !== 'admin') {
-    query.where.userId = req.user.get('id');
-  }
-
-  const dentistInfo = await db.DentistInfo.find(query);
-  if (dentistInfo == null) {
-    return next(new BadRequestError('No dentist info object was found'));
-  }
-    console.log("THE BODY",req.body)
-    const queries = [];
-    const user = req.body.user;
-    const officeInfo = req.body.officeInfo;
-    const officeImages = req.body.officeImages;
-    const pricing = req.body.pricing;
-    const membership = req.body.officeInfo.memberships;
-    const childMembership = req.body.childMembership;
-    const workingHours = req.body.workingHours;
-    const services = req.body.services;
-
-    if (user) {
+  if (pricing.codes) {
+    // update pricing codes.
+    pricing.codes.forEach((item) => {
       queries.push(
-        req.user.update({
-          firstName: user.firstName,
-          lastName: user.lastName,
-          phone: user.phone,
-          avatar: user.avatar,
-          zipCode: user.zipCode,
-          dentistSpecialtyId: user.dentistSpecialtyId
-        })
-      );
-    }
-
-    // update info itself.
-    if (officeInfo) {
-      queries.push(
-        dentistInfo.update({
-          officeName: officeInfo.officeName,
-          url: officeInfo.url,
-          phone: officeInfo.phone,
-          message: officeInfo.message,
-          address: officeInfo.address,
-          city: officeInfo.city,
-          state: officeInfo.state,
-          zipCode: officeInfo.zipCode,
-          logo: officeInfo.logo,
-          acceptsChildren: officeInfo.acceptsChildren,
-          childStartingAge: officeInfo.childStartingAge,
-          marketplaceOptIn: officeInfo.marketplaceOptIn
-        })
-      );
-    }
-
-    if (pricing.codes) {
-      // update pricing codes.
-      pricing.codes.forEach((item) => {
-        queries.push(
-          db.MembershipItem.update({
-            price: item.price,
-          }, {
-            where: {
-              id: item.id,
-            }
-          }))
-      });
-    }
-
-    if (membership.length > 0) {
-      membership.forEach(m => {
-        queries.push(
-          db.Membership.update({
-            active: false
-          },
-          {
-            where: {
-              dentistInfoId: user.dentistInfo.id,
-              type: m.type,
-              subscription_age_group: m.subscription_age_group
-            }
-          })
-        );
-        queries.push(
-          db.Membership.create({
-            ..._.omit(m, ['id']),
-            dentistInfoId: user.dentistInfo.id,
-            userId: user.id
-          })
-        );
-      });
-    }
-
-    if (user.phone) {
-      queries.push(
-        db.Phone.update({
-          number: user.phone,
+        db.MembershipItem.update({
+          price: item.price,
         }, {
           where: {
-            userId: dentistInfo.get('userId'),
+            id: item.id,
+          }
+        }))
+    });
+  }
+
+  if (pricing) {
+    const currentMemberships = req.locals.dentistInfo.memberships.map(m => m.toJSON());
+    const alteredMemberships = [];
+    for (const key in pricing) {
+      if (key !== "codes") {
+        alteredMemberships.push(pricing[key]);
+      }
+    }
+    currentMemberships.forEach(cm => {
+      const alteredMembership = alteredMemberships.find(am => am.id === cm.id);
+      if (alteredMembership) {
+        if (cm.price !== alteredMembership.value) {
+          shouldRefresh = true;
+          // Disable old membership
+          queries.push(
+            db.Membership.update({
+              active: false
+            },
+            {
+              where: {
+                id: cm.id
+              }
+            })
+          );
+          // Create New membership
+          queries.push(
+            db.Membership.create({
+              name: cm.name,
+              userId: user.id,
+              discount: cm.discount,
+              price: alteredMembership.value,
+              type: cm.type,
+              subscription_age_group: cm.subscription_age_group,
+              dentistInfoId: user.dentistInfo.id,
+              active: true
+            })
+          );
+          notifyPlanUpdate(cm.id, membership.name, membership.price);
+        }
+      }
+    });
+    // Is discount percentage changed?
+    const anyActiveMembership = currentMemberships.find(cm => cm.active);
+    if (parseInt(pricing.treatmentDiscount) !== anyActiveMembership.discount) {
+      shouldRefresh = true;
+      const membershipIdsToUpdate = currentMemberships
+        .filter(cm => cm.active)
+        .map(cm => cm.id);
+      queries.push(
+        db.Membership.update({
+          discount: parseInt(pricing.treatmentDiscount)
+        }, {
+          where: {
+            id: membershipIdsToUpdate
           }
         })
       );
     }
+  }
 
-    if (workingHours) {
-      // update working hours
-      workingHours.forEach((workingHour) => {
-        queries.push(
-          db.WorkingHours.update({
-            isOpen: workingHour.isOpen,
-            startAt: workingHour.startAt,
-            endAt: workingHour.endAt,
-          }, {
-            where: {
-              dentistInfoId: dentistInfo.get('id'),
-              day: workingHour.day,
-            }
-          })
-        );
-      });
-    }
-
-    if (officeInfo.services) {
-      const previousServices = dentistInfo.get('services');
-      // Go through the services to add.
-      for (const service of officeInfo.services) {
-        const serviceAlreadyExists =
-            previousServices.find(s => s.dataValues.serviceId === service.id);
-        if (!serviceAlreadyExists) {
-          queries.push(db.DentistInfoService.upsert({
-            serviceId: service.id,
-            dentistInfoId: dentistInfo.get('id')
-          }));
+  if (user.phone) {
+    queries.push(
+      db.Phone.update({
+        number: user.phone,
+      }, {
+        where: {
+          userId: dentistInfo.get('userId'),
         }
-      }
+      })
+    );
+  }
 
-      // Go through the services to destroy.
-      for (const service of previousServices) {
-        const serviceShouldExist =
-            officeInfo.services.find(s => s.id === service.dataValues.serviceId);
-        if (!serviceShouldExist) {
-          queries.push(db.DentistInfoService.destroy({
-            where: {
-              serviceId: service.dataValues.serviceId,
-              dentistInfoId: dentistInfo.get('id'),
-            }
-          }));
-        }
-      }
-    }
+  if (workingHours) {
+    // update working hours
+    workingHours.forEach((workingHour) => {
+      queries.push(
+        db.WorkingHours.update({
+          isOpen: workingHour.isOpen,
+          startAt: workingHour.startAt,
+          endAt: workingHour.endAt,
+        }, {
+          where: {
+            dentistInfoId: dentistInfo.get('id'),
+            day: workingHour.day,
+          }
+        })
+      );
+    });
+  }
 
-    if (officeInfo.officeImages) {
-      // update office images.
-      const previousImages = dentistInfo.get('officeImages');
+  // Logic to Add / Delete Services offered by the Dentist.
+  if (officeInfo.services) {
+    const originalServices = dentistInfo.get('services').map(s => s.service.id);
+    const alteredServices = officeInfo.services.map(s => s.id);
 
-      // Go through the images to add.
-      officeInfo.officeImages.forEach((imageUrl) => {
-        const imageAlreadyExists =
-            previousImages.find(s => s.dataValues.url === imageUrl);
-        if (!imageAlreadyExists) {
+    const servicesDiff = processDiff(originalServices, alteredServices);
+    if (!servicesDiff.isSame) {
+      // Add services check
+      if (servicesDiff.addedItems.length) {
+        servicesDiff.addedItems.forEach(serviceId => {
           queries.push(
-            db.DentistInfoPhotos.upsert({
-              url: imageUrl,
-              dentistInfoId: dentistInfo.get('id')
+            db.DentistInfoService.create({
+              dentistInfoId: dentistInfo.get('id'),
+              serviceId
             })
           );
-        }
-      });
+        });
+      }
 
-      // Go through the office images to destroy.
-      for (const instance of previousImages) {
-        const imageShouldExist =
-            officeInfo.officeImages.find(s => s.url === instance.dataValues.url);
-        if (!imageShouldExist) {
-          queries.push(db.DentistInfoPhotos.destroy({
-            where: {
-              url: instance.dataValues.url,
-              dentistInfoId: dentistInfo.get('id'),
-            }
-          }));
-        }
+      // Delete services check
+      if (servicesDiff.removedItems.length) {
+        servicesDiff.removedItems.forEach(serviceId => {
+          queries.push(
+            db.DentistInfoService.destroy({
+              where: {
+                dentistInfoId: dentistInfo.get('id'),
+                serviceId
+              }
+            })
+          );
+        });
       }
     }
+  }
+
+  if (officeInfo.officeImages) {
+    // update office images.
+    const previousImages = dentistInfo.get('officeImages');
+
+    // Go through the images to add.
+    officeInfo.officeImages.forEach((imageUrl) => {
+      const imageAlreadyExists =
+          previousImages.find(s => s.dataValues.url === imageUrl);
+      if (!imageAlreadyExists) {
+        queries.push(
+          db.DentistInfoPhotos.upsert({
+            url: imageUrl,
+            dentistInfoId: dentistInfo.get('id')
+          })
+        );
+      }
+    });
+
+    // Go through the office images to destroy.
+    for (const instance of previousImages) {
+      const imageShouldExist =
+          officeInfo.officeImages.find(s => s.url === instance.dataValues.url);
+      if (!imageShouldExist) {
+        queries.push(db.DentistInfoPhotos.destroy({
+          where: {
+            url: instance.dataValues.url,
+            dentistInfoId: dentistInfo.get('id'),
+          }
+        }));
+      }
+    }
+  }
   Promise.all(queries).then(data => {
-    return;
+    return res.status(200).send({ shouldRefresh });
   },err => {
     console.log("Error in dentist update");
     console.log(err);
+    return res.status(500).send(err);
   });
-  res.json({});
 }
 
 /**
@@ -340,22 +374,24 @@ function deleteDentistInfoPhoto (req, res, next) {
 const router = new Router({ mergeParams: true });
 
 router
+  .route('/:dentistId/edit/:dentistInfoId')
+  .post(
+    userRequired,
+    injectDentistInfo('dentistId'),
+    updateDentistInfo);
+
+router
   .route('/')
   .get(
     userRequired,
     injectDentistInfo(),
-    getDentistInfo);
+    getDentistInfo,
+    getCustomMembership);
 
 router
   .route('/:dentistInfoId/photos/:dentistInfoPhotoId')
   .delete(
     userRequired,
     deleteDentistInfoPhoto);
-
-router
-  .route('/:dentistInfoId')
-  .post(
-    userRequired,
-    updateDentistInfo);
 
 export default router;
