@@ -9,7 +9,7 @@ import Moment from 'moment';
 import db from '../../models';
 import stripe from '../stripe';
 import { BadRequestError } from '../errors';
-import { reenrollMember } from '../../utils/reenroll';
+import { reenrollMember, performEnrollmentWithoutProration } from '../../utils/reenroll';
 import { changePlanUtil } from '../../utils/change_plan';
 
 // ────────────────────────────────────────────────────────────────────────────────
@@ -136,17 +136,92 @@ function subscribe(req, res, next) {
  * @param {Object} res - the express response
  * @param {Function} next - the next middleware function
  */
-function reEnroll(req, res) {
-  const memberId = req.params.userId;
-  const membershipId = req.query.membershipId;
-  const currentUserId = req.user.get('id');
+async function reEnroll(req, res) {
+  try {
+    const memberId = req.params.userId;
+    const membershipId = req.query.membershipId;
+    const currentUserId = req.user.get('id');
 
-  reenrollMember(memberId, currentUserId, membershipId).then(subscription => {
-    console.log(145,subscription);
-    res.status(200).send({ data: subscription });
-  }, err => {
-    res.status(500).send(err);
-  });
+    let user = await db.User.findOne({ where: { id: memberId } });
+    if (user.addedBy) {
+      user = await db.User.findOne({ where: { id: user.addedBy } });
+    }
+    const paymentProfile = await db.PaymentProfile.findOne({ where: { primaryAccountHolder: user.id } });
+    const accountHolderSubscriptions = await db.Subscription.findAll({
+      where: {
+        paymentProfileId: paymentProfile.id
+      },
+      include: [{
+        model: db.Membership,
+        as: 'membership',
+      }]
+    });
+    const membershipPlan = await db.Membership.findOne({ where: { id: membershipId } });
+    const userSubscription = await db.Subscription.findOne({ 
+      where: {
+        clientId: memberId,
+        dentistId: membershipPlan.userId
+      },
+      include: [{
+        model: db.Membership,
+        as: 'membership',
+      }]
+    });
+    const {
+      stripeSubscriptionId,
+      stripeSubscriptionItemId,
+      status
+    } = userSubscription;
+
+    // throw an error if the subscription is active
+    if (stripeSubscriptionId || stripeSubscriptionItemId || status === 'active') {
+      return res.status(400).send({ errors: "User already has an active subscription" });
+    }
+
+    const perform = () => {
+      return new Promise((resolve, reject) => {
+        performEnrollmentWithoutProration(accountHolderSubscriptions, membershipPlan, userSubscription, (err, data) => {
+          if (err) {
+            return reject(err);
+          }
+          return resolve(data);
+        });
+      });
+    };
+
+    await perform();
+    // Do not fire penality if user is enrolling for the first time.
+    // We identify this if the stripeSubscriptionIdUpdatedAt is null.
+    const isEnrollingFirstTime = userSubscription.stripeSubscriptionIdUpdatedAt ? false : true;
+    if (user.reEnrollmentFeeWaiver == true && !isEnrollingFirstTime) {
+      const invoiceItem = await stripe.createInvoiceItem({
+        customer: paymentProfile.stripeCustomerId,
+        amount: RE_ENROLLMENT_PENALTY,
+        currency: 'usd',
+        description: 're-enrollment Fee'
+      });
+      console.log("Invoice item for reenrollOperation success for user Id -> " + paymentProfile.primaryAccountHolder);
+      await db.Penality.create({
+        clientId: userSubscription.clientId,
+        dentistId: userSubscription.dentistId,
+        type: 'reenrollment',
+        amount: 9900
+      });
+    }
+    const subscription = await db.Subscription.findOne({
+      where: {
+        id: userSubscription.id
+      },
+      include: [{
+        model: db.Membership,
+        as: 'membership',
+      }]
+    });
+    return res.status(200).send({ data: subscription });
+  } catch(e) {
+    console.log(e);
+    return res.status(500).send({ message: "Internal Server Error" });
+  }
 }
 
 /**
@@ -269,6 +344,12 @@ async function cancelSubscription(req, res) {
     ) {
       console.log("Cancellation Penality charge Issued successfully for user -> "+primaryUser.firstName+" "+primaryUser.lastName);
       const issueCharge = await stripe.issueCharge(EARLY_CANCELLATION_PENALTY, paymentProfile.stripeCustomerId, 'Early Cancellation Penalty Charge');
+      await db.Penality.create({
+        clientId: subscription.clientId,
+        dentistId: subscription.dentistId,
+        type: 'cancellation',
+        amount: EARLY_CANCELLATION_PENALTY
+      });
     }
 
     const stripeSubscription = await stripe.getSubscription(subscription.stripeSubscriptionId);
@@ -288,10 +369,11 @@ async function cancelSubscription(req, res) {
       const subItemQuantity = subscriptionItem.quantity;
       if (subItemQuantity > 1) {
         await stripe.updateSubscriptionItem(subscriptionItem.id, {
-          quantity: subscriptionItem.quantity - 1
+          quantity: subscriptionItem.quantity - 1,
+          prorate: false
         });
       } else {
-        await stripe.deleteSubscriptionItem(subscriptionItem.id);
+        await stripe.deleteSubscriptionItem(subscriptionItem.id, { prorate: false });
       }
     }
 
