@@ -1,46 +1,24 @@
-// export default function (sequelize, DataTypes) {
-//   const Refunds = DataTypes.define('Refunds', {
-//     transId: {
-//       type: DataTypes.STRING,
-//       allowNull: false,
-//     },
-//     authorizeId: {
-//       type: DataTypes.STRING,
-//       allowNull: false,
-//     },
-//     amount: {
-//       type: DataTypes.FLOAT,
-//       defaultValue: 0,
-//       allowNull: false,
-//     },
-//   }, {
-//     tableName: 'refunds',
-//     timestamps: true,
-//   });
-
-//   return Refunds;
-// }
-
 
 // ────────────────────────────────────────────────────────────────────────────────
 // MODEL
 
 export default function (Sequelize, DataTypes) {
-  const Refund = Sequelize.define('Refund', {
+  const Payment = Sequelize.define('Payment', {
     clientId: {
       type: DataTypes.INTEGER,
       allowNull: false,
-      validate: { notEmpty: true },
     },
     dentistId: {
       type: DataTypes.INTEGER,
       allowNull: false,
-      validate: { notEmpty: true },
+    },
+    stripeSubscriptionId: {
+      type: DataTypes.STRING,
+      allowNull: false,
     },
     amount: {
       type: DataTypes.INTEGER,
       allowNull: false,
-      validate: { notEmpty: true },
     },
     createdAt: {
       type: DataTypes.DATE,
@@ -53,9 +31,9 @@ export default function (Sequelize, DataTypes) {
       defaultValue: Sequelize.fn('NOW'),
     },
   }, {
-    tableName: 'refunds',
+    tableName: 'payments',
     timestamps: false,
 });
 
-  return Refund;
+  return Payment;
 }
